Add znidomb to iset.mm (#4981)
* Add znidom to iset.mm

We don't have Field in iset.mm currently.  But IDomn is not hard.  The
proof is a slightly modified version of the set.mm proof of znfld .

* Add znidomb to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but needs
various small changes in various places.  Don't mention fields in the
comment as we haven't proved anything about fields yet in iset.mm.

* Add znrrg to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with only two small
changes of not equal to apart.