Add "m" to iset.mm conventions and a few related cleanups (#4980)
* add "m" to iset.mm conventions

This is a long-standing convention which has somehow eluded
documentation until now.

* Add r19.3rzv and r19.28zv to mmil.html

As far as I noticed, similar theorems which were intuitionized around
the same time are already listed in mmil.html.

* Remove mpt2 from conventions

We completed the rename of mpt2 to mpo a while ago, so there is no
longer a need to mention the older convention.

* Remove rgenm from iset.mm

In the words of Norm when he approved the removal of rgenz from
set.mm on July 23, 2021, "Yes, [rgenz] should be removed. (I can't
remember what I intended it for, and I don't see any uses for it.)"

rgenm readily follows from other theorems like rgen and 19.8a so
there shouldn't be any difficulty if a proof does want similar
logic.

rgenm is unused.