add a few more gsum theorems to iset.mm (#4982)
* copy gsumvallem2 from set.mm to iset.mm

* copy gsumsubm from set.mm to iset.mm

* Add gsumfzval to iset.mm

This is a tidy expression for gsum which should be much easier to deal
with than igsumval directly.

Also, wrap igsumval in ${ $} without otherwise changing it.

* Add gsumfzz to iset.mm

This is gsumz from set.mm but where the sum is indexed by consecutive
integers.  The proof is via gsumfzval and induction.

* add gsum/Word theorems to mmil.html

This is gsumwsubmcl , gsumws1 , gsumwcl , gsumsgrpccat ,
gsumccat , gsumws2 , gsumccatsn , gsumspl , gsumwmhm ,
and gsumwspan

* add gsum/support theorems to mmil.html

This is gsumval3a , gsumval3eu , gsumval3lem1 , gsumval3lem2 ,
gsumval3 , gsumcllem , gsumzres , gsumzcl2 , gsumzcl ,
gsumzf1o , gsumres , and gsumcl2

* Add gsumfzcl to iset.mm

This is gsumcl from set.mm except only for the case where the sum is
indexed by consecutive integers, not for any finitely supported set.
The proof is a new one.

* Add gsumf1o to mmil.html