add more Word/gsum theorems to iset.mm (#4988)
* Add gsumwsubmcl to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but it
needs changes to show decidability and also for differences in gsum and
seq theorems.

* copy gsumwcl from set.mm to iset.mm

* Add seq1g to iset.mm

This is seq1 from set.mm with two set existence hypotheses.  In the case
where F and .+ are sets this is much simpler than using seq3-1 directly.
The proof is via seq3-1 and is not difficult but it is a handful of
steps.

* Add seqp1g to iset.mm

This is seqp1 from set.mm with two additional set existence conditions.
The proof uses seq3p1

* Add seqclg to iset.mm

This is seqcl from set.mm with additional set existence conditions.
The proof is seq3clss plus a handful of steps.

* Add seqhomog to iset.mm

This is seqhomo from set.mm with added set existence hypotheses.  The
proof is the set.mm proof except that it needs small changes in a number
of places to apply the set existence hypotheses.

* Add gsumwmhm to iset.mm

Stated as in set.mm.  The proof is basically the same as in set.mm but
it needs changes in several places for differences in theorems about
finite sets, gsum , and seq .