add some basel prerequisites and notes in iset.mm (#4992)
* add fnfvof to mmil.html

* Add ofc1g and ofc2g to iset.mm

This is ofc1 and ofc2 from set.mm with an additional hypotheses
similar to the one in ofvalg .  The proofs are similar to the set.mm proofs.

* add caofass to mmil.html

* Add caofdig to iset.mm

This is caofdi from set.mm with set existence conditions added.  The
proof is similar to the set.mm proof.

* add caofdir to mmil.html

* add caonncan to mmil.html

* Add ofnegsub to iset.mm

Stated as in set.mm.  The proof is the set.mm proof but with changes in
various places for differences in oF theorems.

* Add irrmulap to iset.mm

This is irrmul but with irrational in place of not rational.  The proof
follows without much difficulty from existing theorems including apdivmuld .

* Add sinltxirr to iset.mm

This is sinltx from set.mm but just for irrational numbers.  The proof
is by considering the A < 1 and 1 < A cases in turn.

* add df-ply to mmil.html

* add df-idp to mmil.html

* add df-coe to mmil.html

* Add df-dgr to mmil.html

* Add vieta1 to mmil.html

* Add fta1 to mmil.html

* Add notes about basel to mmil.html

This is basellem1 , basellem2 , basellem3 , basellem4 ,
basellem5 , basellem6b , basellem7 , basellem8 , basellem9 , and
basel