Add seqf1og to iset.mm
This is similar to seq3f1o in iset.mm and seqf1o in set.mm but the
conditions are slightly different.  The proof, including lemmas,
is the set.mm proof with changes in various places for set existence,
decidability of various propositions, and differences in seq theorems.

Includes lemmas seqf1oglem2a , seqf1oglem1 , and seqf1oglem2 which
are based on seqf1olem2a , seqf1olem1 , and seqf1olem2 from set.mm.