Remove unneeded hypotheses
This is for plycj and coecj in set.mm.  Because N is not mentioned
in other hypotheses or the conclusion, there is no need for
the N = ( deg ` F ) hypothesis and the proof can just use ( deg ` F )
where it had used N .

The changes to plycjlem are just cosmetic (whitespace etc).