add a few more polynomial theorems to iset.mm (#5021)
* Add plyco to iset.mm

Stated as in set.mm.  The proof, although it is much rearranged, is
based on the set.mm proof.  The "degree" (really an upper bound
for the degree) and coefficients come from elply2 rather than from
deg and coeff .

Parts of the proof are broken out into lemma plycolemc .

* Add plycn to mmil.html

* Remove unneeded hypotheses

This is for plycj and coecj in set.mm.  Because N is not mentioned
in other hypotheses or the conclusion, there is no need for
the N = ( deg ` F ) hypothesis and the proof can just use ( deg ` F )
where it had used N .

The changes to plycjlem are just cosmetic (whitespace etc).

* Add plycj to iset.mm

Stated as in set.mm.  The proof is generally similar but the details are
different because it uses elply in place of deg and coeff .

Includes lemma plycjlemc which is similar to plycjlem from set.mm but
which specifies (an upper bound for a) degree and coefficients
explicitly rather than via deg and coeff .

Add coecj to mmil.html

* Add plyrecj to iset.mm

Stated as in set.mm.  The proof is based on the set.mm proof although it
is changed in many places, especially to handle coefficients via elply
rather than the coeff syntax.