Add dist, le and UnifSet slots to CCfld in iset.mm (#5029)
* Change distinct variables for mpofvex in iset.mm

This is removing the $d B x $. condition for both mpofvex and mpofvexi .
Not only is it not needed (the proof doesn't use it), but there are a
number of places where we have B depend on x (see especially mpoxeldm
and the section containing it, or for an example df-fg ).

Also applies to mpofvexi .

* Add metuex to iset.mm.

This is a set existence theorem of the sort that we need because we
don't have ovex .  The proof is via set existence theorems such as
mpofvexi and mptfvex .

* Add cndsex to iset.mm

This is a simple set existence theorem.  The proof is taken from a
portion of the cntopex proof.  Revise the cntopex proof to refer to
cndsex .

* Change CCfld definition in iset.mm

Copy df-cnfld from set.mm and replace df-icnfld with it.  The only change
is a small one to the comment.

Revise proofs of cnfldstr , cnfldbas , mpocnfldadd , mpocnfldmul ,
cnfldcj and , cnfldtset accordingly.  They need to account for the new
slots but are otherwise similar to the previous iset.mm proofs.

Add cnfldle and cnfldds to iset.mm.  Stated as in set.mm (with the
proof being the set.mm proof, slightly modified).

Update mmil.html accordingly

* copy cnfldms from set.mm to iset.mm

* copy cnfldxms from set.mm to iset.mm

* copy cnfldtps from set.mm to iset.mm

* copy cnfldtopn from set.mm to iset.mm

* copy cnfldtopon from set.mm to iset.mm

* copy cnfldtop from set.mm to iset.mm

* copy unicntop from set.mm to iset.mm

* copy cnopn from set.mm to iset.mm

* copy tgioo2 from set.mm to iset.mm

* copy rerest from set.mm to iset.mm