polynomials are continuous (iset.mm) (#5031)
* Add cncfcn1 to iset.mm

Stated as in set.mm.  The proof is via cncfcn1cntop .

* Add fsumcn to iset.mm

Stated as in set.mm.  The proof is via fsumcncntop .

* add ovmpot to mmil.html

* copy fvmpopr2d from set.mm to iset.mm

* Add mpomulcn to iset.mm

Stated as in set.mm.  The proof is the set.mm proof, with a few small
changes.

* Add expcn to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some changes
around replacing ovmpot with ovmpog .

* Add plycn to iset.mm

Stated as in set.mmm.  The proof is basically the set.mm proof although
many steps have to change to use elply in place of coeff and deg .