Add dvmptfsum to iset.mm
Stated as in set.mm.  The proof is based on the set.mm proof but
needs changes in many places.  Some of them are just rearranging
the logic so that finiteness is available (rather than deduced
from ssfi ).  Some of them are to use dvconstss in place of
derivative restriction theorems.