A revised attempt at ax-mulf comment wording (#5032)
* Clarify ax-mulf comment

The ax-mulf comment is confusing.
Its text "is not a bona fide axiom for complex numbers"
makes it sound like this statement can't be an axiom or that
Metamath can't handle such axioms, which is obviously untrue.
Also, the cited paper doesn't justify this axiom.

This commit attempts to clarify the comment for ax-mulf
in the hopes of making it clearer.

* Revise wording of ax-mulf and axmulf

This is in response to discussion on github and includes:
- focus on the consequences of using it or not, with less language about
  which we prefer (because there doesn't seem to be a consensus about
  that).
- compare with other operations
- refer to mulcl rather than ax-mulcl directly
- mention mpomulf

---------

Co-authored-by: David A. Wheeler <dwheeler@dwheeler.com> (an earlier draft than what is being merged here)