the derivative of a polynomial (iset.mm) (#5042)
* Add dvidrelem to iset.mm

Analogue of ~ dvidlemap for real numbers rather than complex numbers.
The proof is similar to the proof of dvidlemap although most steps need
to be adjusted to use real numbers and the topology on the real numbers.

* add dvconstre to iset.mm

This is like dvconst but for real derivatives rather than complex
derivatives.  The proof is similar to the dvconst proof in iset.mm.

* Add dvidre to iset.mm

This is dvid but for real numbers rather than complex numbers.  The
proof is similar to the dvid proof.

* Add dvmptid to iset.mm

The proof is similar to the dvmptidcn proof and uses dvid and
dvidre .

* Add dvmptc to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof in some
ways but uses dvconst and dvconstre .

* Add dvidsslem to iset.mm

Similar to dvidlemap but where the function doesn't need to be defined
on all of CC .  The proof is similar to the proof of dvidlemap or
dvidrelem .

* Add dvconstss to iset.mm

This is similar to dvconst and dvconstre but for a function defined on
an open set.

* Add dvmptfsum to iset.mm

Stated as in set.mm.  The proof is based on the set.mm proof but
needs changes in many places.  Some of them are just rearranging
the logic so that finiteness is available (rather than deduced
from ssfi ).  Some of them are to use dvconstss in place of
derivative restriction theorems.

* Add dvply1 to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof, although
various parts need changes for various reasons.

* Adjust dvidrelem and dvidsslem comments

Since the names end in lem seems logical to use the "Lemma for"
convention.