Add sigma (df-sgm) to iset.mm (#5049)
* add positive divisors function to iset.mm

This is the sigma syntax and df-sgm .  Copied without change from
set.mm.

* Add dvdsfi to iset.mm

A natural number has finitely many divisors.  The proof is taken
from a portion of the phisum proof.  Shorten the phisum proof using
dvdsfi .

* Add sgmval to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes
needed to use fsumcl instead of sumex .

* Add sgmval2 to iset.mm

Stated as in set.mm.  The proof is the set.mm with a small change for
differences in ^c theorems.

* Add 0sgm to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes
related to differences in finite set theorems.

* Add sgmf to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some small
changes for differences in finiteness and ^c theorems.

* copy sgmcl from set.mm to iset.mm

* Add sgmnncl to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes
around theorems related to finiteness and nonempty sets.

* copy dvdsppwf1o from set.mm to iset.mm

* Add rpcxpmul2 to iset.mm

This is cxpmul2 from set.mm but where the base has to be a positive
real, not any complex number.  The proof is based on a portion of the
set.mm proof although most steps need some changes.

* Add rpcxpmul2d to iset.mm

Deduction form of rpcxpmul2 .

* Add sgmppw to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some small
changes related to set existence and differences in ^c theorems.

* Add 0sgmppw to iset.mm

Stated as in set.mm.  The proof is the set.mm with a few small changes
to account for differences in finite set theorems.

* Add 1sgmprm to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a small change
for differences in ^c theorems.

* Add 1sgm2ppw to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with several
changes for differences in ^c theorems and changing not equal to
apart.

* Add mpodvdsmulf1o to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a few changes
for set existence and apartness vs not equal.

* Add fsumdvdsmul to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes for
finite set theorems and set existence.  In some parts of the proof the
changes are significant but some parts of the proof, and the overall
logic, are unchanged.

* Add sgmmul to iset.mm

Stated as in set.mm.  The proof is the set.mm with small changes for
differences in ^c theorems.

* Add dvdsfi to set.mm

Although the proof is shorter than in iset.mm it seems like this may be
useful.

Use it to shorten several proofs.

* Revise dvdsppwf1o comment in set.mm and iset.mm

It isn't clear what "prime count" was supposed to mean here; hopefully
"exponent" (of the prime power) is clearer.

* Adjust cxpmul2 comments

The change is to say "nonnegative integer" rather than "integer".  The
theorem had already said NN0 rather than ZZ .

This is cxpmul2 and cxpmul2d in set.mm and rpcxpmul2 and rpcxpmul2d in
iset.mm.