Intuitionize section "Decimal arithmetic (cont.)" (#5052)
* copy dec2dvds from set.mm to iset.mm

* copy dec5dvds from set.mm to iset.mm

* Rename addid2i to addlidi in iset.mm

This is to match set.mm.

* copy dec5dvds2 from set.mm to iset.mm

* copy dec5nprm from set.mm to iset.mm

* copy dec2nprm from set.mm to iset.mm

* Add modxai to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some small
changes for differences in modulus theorems.

* copy mod2xi from set.mm to iset.mm

* copy modxp1i from set.mm to iset.mm

* copy mod2xnegi from set.mm to iset.mm

* Add modsubi to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with minor changes
for differences in modulus theorems.

* copy gcdi from set.mm to iset.mm

* copy gcdmodi from set.mm to iset.mm

* Rename addid1i to addridi in iset.mm

This is to match current set.mm naming.

* copy decexp2 from set.mm to iset.mm

* copy numexp0 from set.mm to iset.mm

* copy numexp1 from set.mm to iset.mm

* copy numexpp1 from set.mm to iset.mm

* copy numexp2x from set.mm to iset.mm

* Rename mulid1i to mulridi in iset.mm

This is to match current set.mm naming.

* copy decsplit0b from set.mm to iset.mm

* copy decsplit0 from set.mm to iset.mm

* copy decsplit1 from set.mm to iset.mm

* copy decsplit from set.mm to iset.mm

* copy karatsuba from set.mm to iset.mm

* copy 2exp4 from set.mm to iset.mm

* copy 2exp5 from set.mm to iset.mm

* copy 2exp6 from set.mm to iset.mm

* Add 2exp7 to iset.mm

Stated as in set.mm.  The proof had to be loaded and saved in mmj2 but
is otherwise unchanged from the set.mm proof.

* copy 2exp8 from set.mm to iset.mm

* copy 2exp11 from set.mm to iset.mm

* copy 2exp16 from set.mm to iset.mm

* copy 3exp3 from set.mm to iset.mm

* copy 2expltfac from set.mm to iset.mm

* Fix comments for dec5nprm and dec2nprm .

This is in set.mm and iset.mm and the change is to make the comment
better reflect what the theorem says.

* Delete decexp2 from set.mm and iset.mm

* Remove mod2xnegi from iset.mm

We don't need it yet and the name seems off (what is negative here?) or
at least unexplained.