Intuitionize section Perfect Number Theorem (#5060)
* Add ltexp2d to iset.mm

Deduction form of ltexp2 which iset.mm has.  Mostly the same as the
set.mm ltexp2d but a bit later in the file.

* Add mersenne to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes in
various places for differences between set.mm and iset.mm.

* copy perfect1 from set.mm to iset.mm

* copy perfectlem1 from set.mm to iset.mm

* Add prfidceq to iset.mm

This is a variation on prfidisj .  The proof is from prfidisj and
snfig .

* Allow classes in dcun in iset.mm

The proof doesn't need to change substantively as all the steps in the
proof work with classes in place of setvars.  Update the one proof
which had been using dcun - sumsplitdc - in a similar fashion.

* Add tpfidceq to iset.mm

Similar to prfidcdq but for triples instead of pairs.  The proof from
prfidceq and other theorems is not especially difficult.

* Add perfectlem2 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof but needs changes in
a number of parts for various issues.

* Add perfect to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with one small
change for not equal versus apart.

Add to mm_100.html and update mmil.html.

* Also mention ivthdich in mm_100.html

Not sure there is one iset.mm theorem which fully expresses the
Intermediate Value Theorem situation but this seems as worthy
of mention as ivthinc .

* Shorten perfectlem2 comment

This is in set.mm and iset.mm

* Mention ~ mersenne in ~ perfect comment.